Split up the no arg functions into relevant files
